Rename and optimize property subscriber
