+++ Resources/services.xml
@@ -9,16 +9,15 @@
         <service id="neti_tool_kit.subscriber.global_data" class="NetiToolKit\Subscriber\GlobalData">
             <argument type="service" id="neti_foundation.plugin_manager_config"/>
             <argument type="service" id="session"/>
-            <argument type="service" id="models" />
+            <argument type="service" id="models"/>
             <tag name="shopware.event_subscriber"/>
         </service>
 
-        <service id="neti_tool_kit.subscriber.listing_properties" class="NetiToolKit\Subscriber\ListingProperties">
+        <service id="neti_tool_kit.subscriber.listing_properties" class="NetiToolKit\Subscriber\FrontendProperties">
             <argument type="service" id="shopware_storefront.context_service"/>
             <argument type="service" id="shopware_storefront.property_service"/>
             <argument type="service" id="legacy_struct_converter"/>
             <argument type="service" id="neti_foundation.plugin_manager_config"/>
-            <argument type="service" id="modules" />
             <tag name="shopware.event_subscriber"/>
         </service>
